- cleanup (extracting nested classes, renaming classes, addition of subpackages, renaming packages)
