+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/common/widgets/lists/DummyElement.java
@@ -1,0 +1,14 @@
+package org.phoenicis.javafx.views.common.widgets.lists;
+
+import javafx.scene.layout.VBox;
+
+/**
+ * Created by marc on 15.05.17.
+ */
+public class DummyElement extends VBox {
+    public DummyElement() {
+        super();
+
+        this.getStyleClass().add("iconListCell");
+    }
+}
