+++ phoenicis-javafx/src/main/resources/org/phoenicis/javafx/views/common/widgets/lists/wineMiniature.png
